Embedding-models/pyproject.toml
[project]
name = "embedding-models"
version = "0.1.0"
description = ""
authors = [
    {name = "LaljiSankhat03",email = "laljis@webelight.co.in"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "nltk (>=3.9.2,<4.0.0)",
    "gensim (>=4.4.0,<5.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "tensorflow-hub (>=0.16.1,<0.17.0)",
    "transformers (>=4.57.3,<5.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "google-genai (>=1.57.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "openl3 (>=0.4.2,<0.5.0)",
    "torchcodec (>=0.9.1,<0.10.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "sounddevice (>=0.5.3,<0.6.0)",
    "torchaudio (>=2.9.1,<3.0.0)",
    "speechbrain (>=1.0.3,<2.0.0)",
    "noisereduce (>=3.0.3,<4.0.0)",
]

[tool.poetry]
packages = [{include = "embedding_models", from = "src"}]

[tool.poetry.dependencies]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
